+++ packages/client/src/rpc/createClient.ts
@@ -12,6 +12,7 @@ import {
 import { SignalServerClient } from '../gen/video/sfu/signal_rpc/signal.client';
 import { Logger, LogLevel } from '../coordinator/connection/types';
 import type { Trace } from '../stats';
+import type { SfuResponseWithError } from './retryable';
 
 const defaultOptions: TwirpOptions = {
   baseUrl: '',
@@ -49,16 +50,13 @@ export const withRequestLogger = (
       input: object,
       options: RpcOptions,
     ): UnaryCall => {
-      let invocation: UnaryCall | undefined;
-      try {
-        invocation = next(method, input, options);
-      } finally {
-        logger(level, `Invoked SFU RPC method ${method.name}`, {
-          request: invocation?.request,
-          headers: invocation?.requestHeaders,
-          response: invocation?.response,
-        });
-      }
+      const invocation = next(method, input, options);
+      logger(level, `Invoked SFU RPC method ${method.name}`, {
+        request: invocation.request,
+        headers: invocation.requestHeaders,
+        response: invocation.response,
+      });
+
       return invocation;
     },
   };
@@ -68,6 +66,9 @@ export const withRequestTracer = (trace: Trace): RpcInterceptor => {
   type RpcMethodNames = {
     [K in keyof SignalServerClient as Capitalize<K>]: boolean;
   };
+
+  const traceError = (name: string, input: object, err: unknown) =>
+    trace(`${name}OnFailure`, [err, input]);
 
   const exclusions: Record<string, boolean | undefined> = {
     SendStats: true,
@@ -82,13 +83,17 @@ export const withRequestTracer = (trace: Trace): RpcInterceptor => {
       if (exclusions[method.name as keyof RpcMethodNames]) {
         return next(method, input, options);
       }
-      try {
-        trace(method.name, input);
-        return next(method, input, options);
-      } catch (err) {
-        trace(`${method.name}OnFailure`, [input, err]);
-        throw err;
-      }
+
+      trace(method.name, input);
+      const unaryCall = next(method, input, options);
+      unaryCall.then(
+        (invocation) => {
+          const err = (invocation.response as SfuResponseWithError)?.error;
+          if (err) traceError(method.name, input, err);
+        },
+        (err) => traceError(method.name, input, err),
+      );
+      return unaryCall;
     },
   };
 };