+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,55 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.18.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.17.1...@stream-io/video-react-sdk-1.18.0) (2025-05-22)
+
+### Dependency Updates
+
+- `@stream-io/video-styling` updated to version `1.2.0`
+- `@stream-io/video-client` updated to version `1.23.2`
+- `@stream-io/video-react-bindings` updated to version `1.6.5`
+
+### Features
+
+- add backstage mode for LivestreamPlayer ([#1796](https://github.com/GetStream/stream-video-js/issues/1796)) ([94e2863](https://github.com/GetStream/stream-video-js/commit/94e2863aa8c52542f2ecb53f71485d5d5f2e79fd))
+
+### Bug Fixes
+
+- sequential uPDP ([#1802](https://github.com/GetStream/stream-video-js/issues/1802)) ([f613b97](https://github.com/GetStream/stream-video-js/commit/f613b97ea03e85fd64ca5de76f4e34d1507fef27))
+
+## [1.17.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.17.0...@stream-io/video-react-sdk-1.17.1) (2025-05-21)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.23.1`
+- `@stream-io/video-react-bindings` updated to version `1.6.4`
+
+## [1.17.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.16.2...@stream-io/video-react-sdk-1.17.0) (2025-05-20)
+
+### Dependency Updates
+
+- `@stream-io/audio-filters-web` updated to version `0.4.0`
+- `@stream-io/video-client` updated to version `1.23.0`
+- `@stream-io/video-react-bindings` updated to version `1.6.3`
+
+### Features
+
+- **web:** improved noise cancellation ([#1794](https://github.com/GetStream/stream-video-js/issues/1794)) ([d59f19b](https://github.com/GetStream/stream-video-js/commit/d59f19b1ba1ff83fe5f024d783b868f4e98d3380))
+
+## [1.16.2](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.16.1...@stream-io/video-react-sdk-1.16.2) (2025-05-15)
+
+### Dependency Updates
+
+- `@stream-io/audio-filters-web` updated to version `0.3.1`
+- `@stream-io/video-styling` updated to version `1.1.5`
+- `@stream-io/video-client` updated to version `1.22.2`
+- `@stream-io/video-filters-web` updated to version `0.2.1`
+- `@stream-io/video-react-bindings` updated to version `1.6.2`
+
+### Bug Fixes
+
+- enable chore releases ([#1792](https://github.com/GetStream/stream-video-js/issues/1792)) ([6046654](https://github.com/GetStream/stream-video-js/commit/6046654fe19505a1c115a4fb838759d010540614))
+
 ## [1.16.1](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-1.16.0...@stream-io/video-react-sdk-1.16.1) (2025-05-14)
 
 ### Dependency Updates