+++ packages/styling/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.2.0](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-1.1.5...@stream-io/video-styling-1.2.0) (2025-05-22)
+
+### Features
+
+- add backstage mode for LivestreamPlayer ([#1796](https://github.com/GetStream/stream-video-js/issues/1796)) ([94e2863](https://github.com/GetStream/stream-video-js/commit/94e2863aa8c52542f2ecb53f71485d5d5f2e79fd))
+
+## [1.1.5](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-1.1.4...@stream-io/video-styling-1.1.5) (2025-05-15)
+
+### Bug Fixes
+
+- enable chore releases ([#1792](https://github.com/GetStream/stream-video-js/issues/1792)) ([6046654](https://github.com/GetStream/stream-video-js/commit/6046654fe19505a1c115a4fb838759d010540614))
+
 ## [1.1.4](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-1.1.3...@stream-io/video-styling-1.1.4) (2025-04-15)
 
 ### Bug Fixes